feat: centralize captain utilities
